Break control-auth into multiple files
